+++ pizzashop.Entity/ViewModels/OrderTax.cs
@@ -6,5 +6,5 @@ public class OrderTax
 
     public int Taxid { get; set; }
 
-    public int? Taxvalue { get; set; }
+    public decimal? Taxvalue { get; set; }
 }
